# laravel-json-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TekManager Sistema de Gestão Integrado",
        "description": "Documentação geral da API de acesso aos dados",
        "contact": {
            "email": "cesar@berci.com.br"
        },
        "license": {
            "name": "TekSea ₢ 2024",
            "url": "http://www.teksea.net"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v2/clientes": {
            "get": {
                "summary": "Get all clientes",
                "description": "Tabela que armazena informações dos clientes",
                "operationId": "43ccca345bfd83280de0dc9ac9545e9e",
                "responses": {
                    "200": {
                        "description": "Get all clientes"
                    }
                }
            },
            "post": {
                "summary": "Create a new cliente",
                "operationId": "a57fdcd0f539a4c2e936586ccc8b1a42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Identificador único de cada cliente",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cnpj_cpf": {
                                        "description": "CNPJ ou CPF do cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "nome": {
                                        "description": "Nome completo ou razão social do cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "nome_fantasia": {
                                        "description": "Nome fantasia do cliente, se aplicável",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "uf": {
                                        "description": "Unidade federativa (estado) do cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "cidade": {
                                        "description": "Cidade do cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "representante": {
                                        "description": "Representante associado ao cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "situacao": {
                                        "description": "Situação do cliente (ativo/inativo)",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "created_at": {
                                        "description": "Data de criação do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    },
                                    "updated_at": {
                                        "description": "Data de atualização do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new cliente"
                    }
                }
            }
        },
        "/api/v2/clientes/{id}": {
            "get": {
                "summary": "Get a cliente by ID",
                "operationId": "457fc12be6a9ce35faaa8f8268b0137c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get a cliente by ID"
                    }
                }
            },
            "put": {
                "summary": "Update a cliente",
                "operationId": "af175555811824a382b5236705b23df7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Identificador único de cada cliente",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cnpj_cpf": {
                                        "description": "CNPJ ou CPF do cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "nome": {
                                        "description": "Nome completo ou razão social do cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "nome_fantasia": {
                                        "description": "Nome fantasia do cliente, se aplicável",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "uf": {
                                        "description": "Unidade federativa (estado) do cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "cidade": {
                                        "description": "Cidade do cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "representante": {
                                        "description": "Representante associado ao cliente",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "situacao": {
                                        "description": "Situação do cliente (ativo/inativo)",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "created_at": {
                                        "description": "Data de criação do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    },
                                    "updated_at": {
                                        "description": "Data de atualização do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a cliente"
                    }
                }
            },
            "delete": {
                "summary": "Delete a cliente",
                "operationId": "c936998f9436eb6ff37973459f8a868b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete a cliente"
                    }
                }
            }
        },
...
    }
}